+++ src/app/client/widget/homepage-client/homepage-client.component.ts
@@ -4,6 +4,7 @@ import { Menu } from 'src/app/core/models/menu.model';
 import { HttpResponse } from '@angular/common/http';
 import { ClientService } from 'src/app/core/services/client.service';
 import { SessionStorageService } from 'src/app/core/services/session-storage.service';
+import { ToastrService } from 'ngx-toastr';
 
 @Component({
   selector: 'app-homepage-client',
@@ -12,16 +13,20 @@ import { SessionStorageService } from 'src/app/core/services/session-storage.ser
 })
 export class HomepageClientComponent implements OnInit {
   restoList: RestaurantModel[] = []
-  restoTest = new RestaurantModel();
   menuTest = new Menu();
-  restaurantID: number = 21225;
 
-  constructor(private sessionStorageService: SessionStorageService,private clientService: ClientService) { };
+  constructor(private toastr: ToastrService,private sessionStorageService: SessionStorageService,private clientService: ClientService) { };
 
   ngOnInit(): void {
     let token = this.sessionStorageService.getItem('token');
-    this.clientService.getRestaurant(token).subscribe((response: HttpResponse<any>) => {
-      console.log(response)
+    this.clientService.getRestaurant(token).subscribe((response: RestaurantModel) => {
+      for(let i = 0;i<response.restaurants.length;i++){
+        console.log(response.restaurants[i])
+        let restaurantInformation = response.restaurants[i];
+        this.restoList.push(restaurantInformation);
+      }
+    }, (error) => {
+      this.toastr.error("Erreur lors de la récupération du token : " + error);
     });
 
     this.restoTest.img = "https://images.bfmtv.com/NUJHUYUkXAYVPZAR888_w9rjrNc=/0x0:1196x1192/1196x0/images/-458880.jpg";