+++ src/app/core/models/menu.model.ts
@@ -2,8 +2,10 @@ export class Menu{
     img !:string;
     name !:string;
     description !: string;
+    totalPrice!:string;
     price !: string;
     id!:any;
+    _id!:any;
     drink!:boolean;
     isMenu!:boolean;
 }