+++ src/app/core/services/apiPaths.ts
@@ -6,7 +6,9 @@ export const API = Object.freeze({
     login: '/auth/login',
 
     logs:'/auth/logs',
-
+    getComponentLog:'/component/getLogs',
+    writeComponentLog:'/component/writeLogs',
+    
     editUser: '/user/edit',
     deleteUser: '/user/delete',
     activateUser: '/user/unsuspend',
@@ -15,10 +17,17 @@ export const API = Object.freeze({
     getUsers: '/user/getAll',
 
     getOrder: '/order/getOrder',
-    getRestaurant:'/restaurant/getAll',
-    getRestaurantDetail:'/retaurant/find',
+    getOrders: '/order/getOrders',
+    getAllOrders: '/order/getAllOrders',
 
-    getOrders: '/order/getOrders'
+    createMenu:'/menu/create',
+    createArticle:'/product/add',
+
+    getOrdersSince:'/restaurant/getOrdersSince',
+    getRestaurantInformation: '/restaurant/findByOwner',
+    getRestaurant:'/restaurant/getAll',
+    getRestaurantDetail:'/restaurant/find',
+
 
 
   });