+++ src/app/delivery/widget/homepage-delivery/homepage-delivery.component.ts
@@ -1,5 +1,6 @@
 import { HttpResponse } from '@angular/common/http';
 import { Component } from '@angular/core';
+import { ToastrService } from 'ngx-toastr';
 import { Delivery } from 'src/app/core/models/delivery.model';
 import { DeliveryService } from 'src/app/core/services/delivery.service';
 import { SessionStorageService } from 'src/app/core/services/session-storage.service';
@@ -13,13 +14,15 @@ export class HomepageDeliveryComponent {
   deliveryTest = new Delivery();
   deliveries:Delivery[] = [];
 
-  constructor(private sessionStorageService: SessionStorageService,private deliveryService: DeliveryService) { }
+  constructor(private toastr: ToastrService,private sessionStorageService: SessionStorageService,private deliveryService: DeliveryService) { }
 
 
   ngOnInit(){
     let token = this.sessionStorageService.getItem('token');
     this.deliveryService.getOrdersList(token).subscribe((response: HttpResponse<any>) => {
       console.log(response)
+    }, (error) => {
+      this.toastr.error("Erreur lors de la récupération des commandes : " + error);
     });
 
     this.deliveries.push(this.deliveryTest);