+++ src/app/restaurant/widget/add-article/add-article.component.ts
@@ -1,6 +1,9 @@
 import { Component } from '@angular/core';
 import { FormBuilder, FormControl } from '@angular/forms';
 import { Router, ActivatedRoute } from '@angular/router';
+import { ToastrService } from 'ngx-toastr';
+import { MessageModel } from 'src/app/core/models/message.model';
+import { RestaurantService } from 'src/app/core/services/restaurant.service';
 import { SessionStorageService } from 'src/app/core/services/session-storage.service';
 
 @Component({
@@ -10,9 +13,17 @@ import { SessionStorageService } from 'src/app/core/services/session-storage.ser
 })
 export class AddArticleComponent {
   selectedImage: string | ArrayBuffer | null | undefined;
-  type!:string|null;
+  type!: string | null;
+  token!: any;
+  userID!:any;
 
-  constructor(private formBuilder: FormBuilder,private sessionStorageService: SessionStorageService,private router: Router, private route: ActivatedRoute) { }
+  constructor(private formBuilder: FormBuilder,
+    private sessionStorageService: SessionStorageService,
+    private router: Router,
+    private route: ActivatedRoute,
+    private restaurantService: RestaurantService,
+    private toastr: ToastrService
+  ) { }
 
   articleForm = this.formBuilder.group({
     name: new FormControl(),
@@ -22,11 +33,13 @@ export class AddArticleComponent {
   });
 
   ngOnInit(): void {
+    this.userID = this.sessionStorageService.getItem('restaurantID');
+    this.token = this.sessionStorageService.getItem('token');
     this.type = this.sessionStorageService.getItem('type');
-    if(this.type != 'restaurant'){
+    if (this.type != 'restaurant') {
       this.router.navigate([`/error-page`], { relativeTo: this.route });
     }
-    else{
+    else {
       this.type = "restaurant";
     }
   }
@@ -42,7 +55,19 @@ export class AddArticleComponent {
     }
   }
 
-  submitArticle(){
-    console.log(this.articleForm.value);
+  submitArticle() {
+    this.restaurantService.createArticle(this.token,this.userID,this.articleForm.value).subscribe((response: MessageModel) => {
+      if(response.message == "Product added successfully"){
+        this.toastr.success("L'article a été ajouté avec succès");
+        setTimeout(() => {
+          this.router.navigate(['/restaurant']);
+        }, 2000);
+      }
+      else{
+        this.toastr.error("Erreur lors de la création de l'article");
+      }
+    }, (error) => {
+      this.toastr.error("Erreur lors de la création de l'article : " + error);
+    });
   }
 }